Feature/create monobehavior pong game loop

+++ Assets/Scripts/Utils/Variables/Int/IntReference.cs
@@ -7,6 +7,6 @@
         public int constantValue;
         public IntVariable variableValue;
 
-        public float Value => useConstant ? constantValue : variableValue.Value;
+        public int Value => useConstant ? constantValue : variableValue.Value;
     }
 }
